requirements.txt
streamlit
python-dotenv
PyPDF2
faiss-cpu
langchain-community
langchain-text-splitters
langchain-google-genai
